pyproject.toml
[project]
name = "medminds"
version = "0.1.0"
description = "This is a backend for project medminds"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.7",
    "langserve>=0.3.1",
    "markdown>=3.9",
    "pydantic==2.10.6",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "uvicorn>=0.35.0",
]
